Add basic tests for GdbServer RSP packet handling
A StubTarget has been defined. This serves as a concrete implementation
of the ITarget interface which returns a runtime exception on all
method calls. The aim is that this target can be overridden and
specialized to test the behavior of different RSP packets when they
are provided to the GdbServer.

A minor change has been made to the GdbServer itself to make it
more easy to test. Previously the GdbServer only had one public
entry point, GdbServer::rspServer, which continually receives
packets using GdbServer::rspClientRequest until a 'vKill' packet
causes the server to exist. This was not ideal for testing the
behavior of individual packets, so the
GdbServer::rspClientRequest has been made protected so that it
can be exposed publicly by a subclass in the testing code
(see TestGdbServer).

A set of basic tests have been defined which just check that
the GdbServer gives the expected response when it sees certain
packets. These packets don't change any target state so only
a minimal stub target is sufficient. Future tests will require
a more elaborate stub targets.