Windows: Explicitly target the Win32 platform when testing
CMake's behavior changes for the 'Visual Studio 16 2019' generator
compared to previous versions of Visual Studio. The default used to be
to build for the Win32 platform, but going forward the default will be
to match that of the host system.

This changes the tests on Windows to explicitly target Win32 so that the
same ABI is used, regardless of Visual Studio version used. Future
patches will explicitly build and test for x64.