Add dummy target
This is a placeholder that implements all the methods of ITarget for
testing connections to GDB, but does not emulate any particular
hardware.

This also adds a new macro to the compatibility header
EMBDEBUG_VISIBLE_API, to declare a symbol as externally visible. This is
required for the entry point of the dummytarget on Windows, where
symbols are hidden by default.