Fix type warnings and do first audit of types
Since we now support multiple targets with different pointer sizes, with
compilers that detect different issues regarding implicit type
conversion resulting in possible data loss, this does a first audit of
the issues converting uses of int/uint32_t to types which make the most
sense going forward.

This ties the internals of Gdbserver to more closely match the types
which make the most sense. These are to use size_t for server addresses
and lengths, and uint_reg_t and uint_addr_t for target register and
address values respectively. The net result is that the server should
now be warning clean for both 64-bit Linux and 32/64-bit Windows.