Make the register count a target specified value
Previously this was hardcoded to RISCV_NUM_REGS in the GdbServer,
which obviously won't work for other targets.

This also required a fix to RspPacket as the implicit copy
constructors would cause the internal data buffer point to
be copied, but the destructor would cause the same pointer to
be deleted. The solution was to disable the copy constructor
and assignment operators, and only leave move copy/assignment
operations.