Refactor RspPackets to be non-mutable
Currently the GdbServer has one main RspPacket, which is used to hold
the currently processing request, and is mutated to hold the response(s)
that is sent back to the client. A char * is used throughout the server
with C-style processing for these tasks.

This refactors RspPacket so that once one has been created, no method
can modify its data. A const char * is still accessible for some
analysis, but for sending packets, a new packet must be created and
sent.

To aid in this task, two new classes have been created:

1. RspPacketBuilder - an interface for building packets over a number of
   function calls.
2. ByteView - An interface for viewing and comparing slices of
   const char arrays in a more C++ friendly manner.

All methods for modifying packets have been refactored such that they
create new packets.

Currently, packets have a fixed size of 10,000 bytes, but this can be
modified in the future, or made dynamic if required.