Improve target selection CMake
The targets to build system did not achieve what it was intended to, and
as such all targets were being built unconditionally. This changes that
to correctly enable and disable targets on the command line.

Additionally it adds support for a new pseudo-target "none" which
disables building of all targets.