Replace Jenkins testing with GitHub Actions
Previously Embdebug was tested using a series of Jenkinsfiles on an
internal infrastructure. Now that the server is public and the canonical
source is developed on GitHub, replace this testing with GitHub Actions.
This still tests on Linux, macOS and Windows with errors enabled, and
also runs a separate AddressSanitizer build as well as checking code
formatting as part of code review.